Small comment changes
I’m starting to make a few tweaks to this great theme and I want to
spend more time cleaning up bits of existing code to help people
understand what’s really going on under-the-hood, sotospeak.

So, I’m making small changes to the comments that are either formatted
poorly or hard to read when they are full-on sentences.

No actual working code has been changed.