Add proper link post type
Add and enable proper link post type. Added content-link.php -- which is
a reduced content.php -- that now uses the post title to link to the
first URL that is used in the post content. The link is opened in a new
browser tab/window. The function that extracts the first URL from the
post content was borrowed from the twenty fifteen link.

Fixes: #278